split file for train_chain and faster_rcnn
