+++ tests/links_tests/model_tests/faster_rcnn_tests/test_faster_rcnn.py
@@ -5,8 +5,7 @@ import chainer
 from chainer import testing
 from chainer.testing import attr
 
-from chainercv.links.model.faster_rcnn import FasterRCNN
-from chainercv.links.model.faster_rcnn import FasterRCNNTrainChain
+from dummy_faster_rcnn import DummyFasterRCNN
 
 
 def _random_array(xp, shape):
@@ -22,77 +21,6 @@ def _generate_bbox(xp, n, img_size, min_length, max_length):
     y_max = y_min + xp.random.uniform(min_length, max_length, size=(n,))
     bbox = xp.stack((x_min, y_min, x_max, y_max), axis=1).astype(np.float32)
     return bbox
-
-
-class DummyExtractor(chainer.Link):
-
-    def __init__(self, feat_stride):
-        super(DummyExtractor, self).__init__()
-        self.feat_stride = feat_stride
-
-    def __call__(self, x, test=False):
-        _, _, H, W = x.shape
-        return _random_array(
-            self.xp,
-            (1, 8, H // self.feat_stride, W // self.feat_stride))
-
-
-class DummyHead(chainer.Chain):
-
-    def __init__(self, n_class):
-        super(DummyHead, self).__init__()
-        self.n_class = n_class
-
-    def __call__(self, x, rois, roi_indices, test=False):
-        n_roi = len(rois)
-        cls_locs = chainer.Variable(
-            _random_array(self.xp, (n_roi, self.n_class * 4)))
-        # For each bbox, the score for a selected class is
-        # overwhelmingly higher than the scores for the other classes.
-        score_idx = np.random.randint(
-            low=0, high=self.n_class, size=(n_roi,))
-        scores = self.xp.zeros((n_roi, self.n_class), dtype=np.float32)
-        scores[np.arange(n_roi), score_idx] = 100
-        scores = chainer.Variable(scores)
-
-        return cls_locs, scores
-
-
-class DummyRegionProposalNetwork(chainer.Chain):
-
-    def __init__(self, n_anchor_base, n_roi):
-        super(DummyRegionProposalNetwork, self).__init__()
-        self.n_anchor_base = n_anchor_base
-        self.n_roi = n_roi
-
-    def __call__(self, x, img_size, scale, test=False):
-        B, _, H, W = x.shape
-        n_anchor = self.n_anchor_base * H * W
-
-        rpn_locs = _random_array(self.xp, (B, n_anchor, 4))
-        rpn_cls_scores = _random_array(self.xp, (B, n_anchor, 2))
-        rois = _generate_bbox(
-            self.xp, self.n_roi, img_size[::-1], 16, min(img_size))
-        roi_indices = self.xp.zeros((len(rois),), dtype=np.int32)
-        anchor = _generate_bbox(
-            self.xp, n_anchor, img_size[::-1], 16, min(img_size))
-        return (chainer.Variable(rpn_locs),
-                chainer.Variable(rpn_cls_scores), rois, roi_indices, anchor)
-
-
-class DummyFasterRCNN(FasterRCNN):
-
-    def __init__(self, n_anchor_base, feat_stride, n_fg_class, n_roi,
-                 min_size, max_size
-                 ):
-        super(DummyFasterRCNN, self).__init__(
-            DummyExtractor(feat_stride),
-            DummyRegionProposalNetwork(n_anchor_base, n_roi),
-            DummyHead(n_fg_class + 1),
-            mean=np.array([[[100]], [[122.5]], [[145]]]),
-            min_size=min_size,
-            max_size=max_size
-        )
 
 
 class TestFasterRCNN(unittest.TestCase):
@@ -173,44 +101,6 @@ class TestFasterRCNN(unittest.TestCase):
     def test_predict_gpu(self):
         self.link.to_gpu()
         self.check_predict()
-
-
-class TestFasterRCNNTrainChain(unittest.TestCase):
-
-    def setUp(self):
-        self.n_anchor_base = 6
-        self.feat_stride = 4
-        self.n_fg_class = 3
-        self.n_roi = 24
-        self.link = FasterRCNNTrainChain(DummyFasterRCNN(
-            n_anchor_base=self.n_anchor_base,
-            feat_stride=self.feat_stride,
-            n_fg_class=self.n_fg_class,
-            n_roi=self.n_roi,
-            min_size=600,
-            max_size=800,
-        ))
-
-    def check_call(self):
-        xp = self.link.xp
-
-        n_bbox = 3
-        imgs = chainer.Variable(_random_array(xp, (1, 3, 600, 800)))
-        bboxes = chainer.Variable(
-            _generate_bbox(xp, n_bbox, (600, 800), 16, 350)[np.newaxis])
-        labels = xp.random.randint(0, self.n_fg_class + 1, size=(n_bbox,))
-        labels = chainer.Variable(labels[np.newaxis].astype(np.int32))
-        scale = chainer.Variable(xp.array(1.))
-        loss = self.link(imgs, bboxes, labels, scale)
-        self.assertEqual(loss.shape, ())
-
-    def test_call_cpu(self):
-        self.check_call()
-
-    @attr.gpu
-    def test_call_gpu(self):
-        self.link.to_gpu()
-        self.check_call()
 
 
 @testing.parameterize(
